requirements.txt
FastAPI==0.110.0
uvicorn==0.27.1
google-generativeai==0.4.0
pydantic==2.6.3
python-dotenv==1.0.1
pytest==8.0.2
httpx==0.27.0
